Parse Things Added
I went to the Prse website and followed their directions on how to add
stuff. I imported the appropriate things into each activity and put a
line in each onCreate method. I don't think I broke anything, but I
don't exactly know how this all works. If nothing is working anymore,
this is probably the thing to roll back. Also, files were added to the
lib folder, so be aware of that.