Changed variables from private to package and made it so that the app will not continue if a request has been made to add a school and the box has not been filled out
Unfortunately, if you choose to add a school, do not fill out the
resultant box, and then hit continue, the app does not proceed to the
next screen. This is expected behavior; what is not expected is that
the app will not go to the next screen even after a school has been
entered. Need to change code that determines when the “Add a school”
option has been chosen, not being done in the simplest manner at the
moment. This may improve functionality.

Also did a little more with the test cases